+++ plugins/woocommerce-admin/client/navigation/test/utils.js
@@ -7,7 +7,6 @@ import { getAdminLink } from '@woocommerce/wc-admin-settings';
  * Internal dependencies
  */
 import {
-	addHistoryListener,
 	getDefaultMatchExpression,
 	getFullUrl,
 	getMappedItemsCategories,
@@ -277,36 +276,6 @@ describe( 'getFullUrl', () => {
 		expect( getFullUrl( getAdminLink( 'admin.php?page=testpage' ) ) ).toBe(
 			getAdminLink( 'admin.php?page=testpage' )
 		);
-	} );
-} );
-
-describe( 'addHistoryListener', () => {
-	it( 'should add a custom event to the browser pushState', () => {
-		const mockCallback = jest.fn();
-		const removeListener = addHistoryListener( mockCallback );
-		window.history.pushState( {}, 'Test pushState' );
-		window.history.pushState( {}, 'Test pushState 2' );
-
-		expect( mockCallback.mock.calls.length ).toBe( 2 );
-
-		// Check that events are no longer called after removing the listener.
-		removeListener();
-		window.history.pushState( {}, 'Test pushState 3' );
-		expect( mockCallback.mock.calls.length ).toBe( 2 );
-	} );
-
-	it( 'should add a custom event to the browser replaceState', () => {
-		const mockCallback = jest.fn();
-		const removeListener = addHistoryListener( mockCallback );
-		window.history.replaceState( {}, 'Test replaceState' );
-		window.history.replaceState( {}, 'Test replaceState 2' );
-
-		expect( mockCallback.mock.calls.length ).toBe( 2 );
-
-		// Check that events are no longer called after removing the listener.
-		removeListener();
-		window.history.replaceState( {}, 'Test replaceState 3' );
-		expect( mockCallback.mock.calls.length ).toBe( 2 );
 	} );
 } );