+++ plugins/woocommerce-admin/packages/data/src/navigation/test/reducer.js
@@ -9,6 +9,7 @@ const defaultState = {
 	menuItems: [],
 	favorites: [],
 	requesting: {},
+	persistedQuery: {},
 };
 
 describe( 'navigation reducer', () => {