+++ plugins/woocommerce-admin/packages/data/CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+-   Add management of persisted queries to navigation data store.
+
 # 1.1.1
 
 -   Remove usage of wc-admin alias `@woocommerce/wc-admin-settings`.